Iniciando agiota em sala com a Laysa

+++ 12 - Agência Bancária/src/s9/Conta.java
@@ -39,7 +39,8 @@ abstract class Conta {
     public void depositar( double valor){
         this.setSaldo(getSaldo() + valor);
     }
-    public void sacar(double valor){
+
+    public void sacar(int valor){
         if (this.getSaldo() >= valor){
             System.out.println("Sacando R$" + valor + "...");
             this.setSaldo(getSaldo() - valor);
